+++ functions/src/index.ts
@@ -1,7 +1,7 @@
 
-import { sendNewUserNotificationEmail } from "./admin-functions/newUser";
-import { sendNewEvenParticipantNotification } from "./admin-functions/newEventParticipant";
-import { sendPendingValidationRequest } from "./admin-functions/pendingValidationRequestNotification";
+import { sendNewUserNotificationEmail } from "./triggers/newUser";
+import { sendNewEvenParticipantNotification } from "./triggers/newEventParticipant";
+import { sendPendingValidationRequest } from "./triggers/pendingValidationRequestNotification";
 
 export default {
   sendNewEvenParticipantNotification,
